Create README.md
I designed and developed a gaming website that allows users to create a new game and register it in the database. 
A user-friendly interface was provided, enabling users to carry out operations such as modification and deletion on the registered games.
I used programming languages such as C# and the SQL. ASP.NET MVC, Bootstrap, HTML were employed to expedite the development process.